Power Level and Character Integration
The player now steadily loses power over time and their power level is reflected by their speed as well as the tint of the player character mesh.  I have also done an initial implementation of the electricity arc effect between batteries and the player character, but it may need to be tweaked.  This can be done in parallel with any other changes.