Fixed issues with electricity FX
I changed the behavior for the beam end of the electricity effect from an actor component interface variable to a vector variable  and this allowed it to properly set the beam end location when calling the visual effect event on the battery pickups.
In addition, I also modified the Beam Tangent float values so that the electricity would take a more direct path to the player character.